+++ R/plot.R
@@ -65,7 +65,7 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 	if (is.factor(x[[1]]) && any(is.na(levels(x[[1]]))))
 		x = droplevels(x) # https://github.com/r-spatial/stars/issues/339
 
-	if (join_zlim && !is.character(x[[1]])) {
+	if (join_zlim && !is.character(x[[1]]) && is.null(dots$rgb)) {
 		breaks = get_breaks(x, breaks, nbreaks, dots$logz)
 		if (!inherits(breaks, c("POSIXt", "Date")))
 			breaks = as.numeric(breaks)
@@ -85,7 +85,7 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 		dxy = attr(st_dimensions(x), "raster")$dimensions
 		loop = setdiff(names(dim(x)), dxy) # dimension (name) over which we loop, if any
 		x = aperm(x, c(dxy, loop))
-		zlim = if (join_zlim)
+		zlim = if (join_zlim && is.null(dots$rgb))
 				range(unclass(x[[1]]), na.rm = TRUE)
 			else
 				rep(NA_real_, 2)
@@ -103,9 +103,9 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 		if (length(dims) == 2 || dims[3] == 1 || (!is.null(dots$rgb) && is.numeric(dots$rgb))) { ## ONE IMAGE:
 			# set up key region
 			values = structure(x[[1]], dim = NULL) # array -> vector
-			if (! isTRUE(dots$add) && ! is.null(key.pos) && !all(is.na(values)) &&
+			if (! isTRUE(dots$add) && ! is.null(key.pos) && !all(is.na(values)) && is.null(dots$rgb) &&
 					(is.factor(values) || length(unique(na.omit(values))) > 1) &&
-					length(col) > 1 && is.null(dots$rgb) && !is_curvilinear(x)) { # plot key?
+					length(col) > 1 && !is_curvilinear(x)) { # plot key?
 				switch(key.pos,
 					layout(matrix(c(2,1), nrow = 2, ncol = 1), widths = 1, heights = c(1, key.width)),  # 1 bottom
 					layout(matrix(c(1,2), nrow = 1, ncol = 2), widths = c(key.width, 1), heights = 1),  # 2 left